backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
spacy==3.7.4
scikit-learn==1.4.2
sentence-transformers==2.7.0
sqlalchemy==2.0.30
pydantic-settings==2.2.1
python-dotenv==1.0.1
reportlab==4.2.0
aiofiles==23.2.1
numpy==1.26.4